+++ deduce/lookup_structs.py
@@ -9,9 +9,9 @@ from typing import Optional
 
 import docdeid as dd
 from docdeid.tokenizer import Tokenizer
+from docdeid.ds import DsCollection
 
 from deduce.data.lookup.src import all_lists
-from deduce.depr import DeprecatedDsCollection
 from deduce.lookup_struct_loader import (
     load_eponymous_disease_lookup,
     load_first_name_lookup,
@@ -233,21 +233,7 @@ def get_lookup_structs(
         "explicitly triggered with Deduce(build_lookup_structs=True)."
     )
 
-    lookup_structs = DeprecatedDsCollection(
-        deprecated_items={
-            "prefixes": "prefix",
-            "first_names": "first_name",
-            "first_name_exceptions": None,
-            "interfixes": "interfix",
-            "interfix_surnames": "interfix_surname",
-            "surnames": "surname",
-            "surname_exceptions": None,
-            "streets": "street",
-            "placenames": "placename",
-            "hospitals": "hospital",
-            "healthcare_institutions": "healthcare_institution",
-        }
-    )
+    lookup_structs = DsCollection()
 
     base_items = load_raw_itemsets(base_path=lookup_path, subdirs=all_lists)