+++ deduce/pattern/__init__.py
@@ -1,8 +1,0 @@
-# type: ignore
-
-from .name_patient import (
-    PersonFirstNamePattern,
-    PersonInitialFromNamePattern,
-    PersonInitialsPattern,
-    PersonSurnamePattern,
-)