+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## (unreleased)
+
+### Removed
+- the `config_file` keyword, now replaced by `config` which accepts both filenames and dicts
+- old lookup list names, e.g. `prefixes` now replaced by `prefix`
+- annotator types `custom`, `regexp`, `token_pattern`, `dd_token_pattern` and `annotation_context`, all replaced by setting class directly as `annotator_type` 
+- everything in `deduce.pattern`, patient patterns now replaced by `PatientNameAnnotator`
+
 ## 3.0.2 (2023-02-15)
 
 ### Changed